// src/components/Checkbox/Checkbox.d.ts
export interface CheckboxProps
  extends Omit<
    React.InputHTMLAttributes<HTMLInputElement>,
    'onChange' | 'checked' | 'defaultChecked'
  > {
  label?: string;
  error?: boolean;
  helperText?: string;
  indeterminate?: boolean;
  checked?: boolean;
  defaultChecked?: boolean;
  onChange?: (checked: boolean) => void;
}
declare const Checkbox: import('react').ForwardRefExoticComponent<
  CheckboxProps & import('react').RefAttributes<HTMLInputElement>
>;
export { Checkbox };